Realized that having two adapters is dumb. Refactored down to one. Added logging anywhere we have ParseExceptions. Did some formatting cleanup.
