+++ app/src/main/java/mrayer/photohunt/AlbumListAdapter.java
@@ -1,6 +1,7 @@
 package mrayer.photohunt;
 
 import android.content.Context;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,6 +12,7 @@ import android.widget.TextView;
 import com.parse.FindCallback;
 import com.parse.ParseException;
 import com.parse.ParseQuery;
+import com.parse.ParseUser;
 import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
@@ -19,17 +21,17 @@ import java.util.List;
 /**
  * Created by Matthew on 3/25/2016.
  */
-public class AlbumGalleryListAdapter extends BaseAdapter {
+public class AlbumListAdapter extends BaseAdapter {
     private final Context context;
     private List<PhotoHuntAlbum> albums;
 
-    public AlbumGalleryListAdapter(Context context) {
+    public AlbumListAdapter(Context context) {
+        // make sure to call one of the load functions after constructing this object
         this.context = context;
         albums = new ArrayList<PhotoHuntAlbum>();
-        loadObjects();
     }
 
-    public void loadObjects() {
+    public void loadAllAlbums() {
         ParseQuery<PhotoHuntAlbum> query = ParseQuery.getQuery("PhotoHuntAlbum");
         query.orderByAscending("name");
         query.findInBackground(new FindCallback<PhotoHuntAlbum>() {
@@ -43,6 +45,24 @@ public class AlbumGalleryListAdapter extends BaseAdapter {
         });
     }
 
+    public void loadCurrentUserAlbums() {
+        ParseQuery<PhotoHuntAlbum> query = ParseQuery.getQuery("PhotoHuntAlbum");
+        query.orderByAscending("name");
+        query.whereEqualTo("authorId", ParseUser.getCurrentUser().getObjectId());
+        query.findInBackground(new FindCallback<PhotoHuntAlbum>() {
+            public void done(List<PhotoHuntAlbum> objects, ParseException e) {
+                if (e == null) {
+                    albums.clear();
+                    albums.addAll(objects);
+                    notifyDataSetChanged();
+                }
+                else {
+                    Log.d(Constants.AlbumListAdapter_Tag, e.toString());
+                }
+            }
+        });
+    }
+
     @Override public View getView(int position, View view, ViewGroup parent) {
         ViewHolder holder;
         if (view == null) {
@@ -51,7 +71,8 @@ public class AlbumGalleryListAdapter extends BaseAdapter {
             holder.photo = (ImageView) view.findViewById(R.id.photo);
             holder.name = (TextView) view.findViewById(R.id.name);
             view.setTag(holder);
-        } else {
+        }
+        else {
             holder = (ViewHolder) view.getTag();
         }
